iterating

+++ dev/openai_dqn.py
@@ -96,7 +96,7 @@ class Trainer(object):
 
     def getScreen(self):
         screen = self.env.render(mode='rgb_array').transpose((2, 0, 1))
-        screen = np.ascontiguousarray(screen, dtype=np.float32)
+        screen = np.ascontiguousarray(screen, dtype=np.float16)
         return torch.from_numpy(screen).unsqueeze(0).to(self.device)
